Revert "Detect mismatch between host defined in config and env variable (#2549)"
This reverts commit 314766ab1531c2a0182c6259e0f4a7c397fd977c.
