--- acceptance/auth/bundle_and_env/.databrickscfg
@@ -1,5 +1,0 @@
-[DEFAULT]
-host = https://bar.com
-
-[profile_with_matching_host]
-host = $DATABRICKS_HOST
