--- acceptance/auth/bundle_and_env/databricks.yml
@@ -1,11 +1,0 @@
-bundle:
-  name: test-auth
-
-targets:
-  match:
-    default: true
-    workspace:
-      host: $DATABRICKS_HOST
-  not_match:
-    workspace:
-      host: https://foo.com
