--- acceptance/auth/bundle_and_env/output.txt
@@ -1,35 +1,0 @@
-
-=== Bundle commands load bundle configuration with DATABRICKS_HOST defined, validation not OK (env-defined host doesn't match bundle host)
->>> errcode [CLI] bundle validate -t not_match
-Error: cannot resolve bundle auth configuration: config host mismatch: DATABRICKS_HOST is defined as [DATABRICKS_HOST], but CLI configured to use https://foo.com
-
-Name: test-auth
-Target: not_match
-Workspace:
-  Host: https://foo.com
-
-Found 1 error
-
-Exit code: 1
-
-=== Bundle commands load bundle configuration with DATABRICKS_HOST defined, validation OK (env-defined host matches bundle host)
->>> errcode [CLI] bundle validate -t match
-Name: test-auth
-Target: match
-Workspace:
-  Host: [DATABRICKS_HOST]
-  User: [USERNAME]
-  Path: /Workspace/Users/[USERNAME]/.bundle/test-auth/match
-
-Validation OK!
-
-=== Bundle commands load bundle configuration with -p flag with DATABRICKS_HOST defined, host in profile matches bundle host but env-defined host doesn't
->>> errcode [CLI] bundle validate -t match -p profile_with_matching_host
-Name: test-auth
-Target: match
-Workspace:
-  Host: [DATABRICKS_HOST]
-  User: [USERNAME]
-  Path: /Workspace/Users/[USERNAME]/.bundle/test-auth/match
-
-Validation OK!
