--- acceptance/auth/bundle_and_env/script
@@ -1,15 +1,0 @@
-# Replace placeholder with an actual host URL
-envsubst < databricks.yml > out.yml && mv out.yml databricks.yml
-envsubst < .databrickscfg > out && mv out .databrickscfg
-
-export DATABRICKS_CONFIG_FILE=.databrickscfg
-
-title "Bundle commands load bundle configuration with DATABRICKS_HOST defined, validation not OK (env-defined host doesn't match bundle host)"
-trace errcode $CLI bundle validate -t not_match
-
-title "Bundle commands load bundle configuration with DATABRICKS_HOST defined, validation OK (env-defined host matches bundle host)"
-trace errcode $CLI bundle validate -t match
-
-export DATABRICKS_HOST="https://baz.com"
-title "Bundle commands load bundle configuration with -p flag with DATABRICKS_HOST defined, host in profile matches bundle host but env-defined host doesn't"
-trace errcode $CLI bundle validate -t match -p profile_with_matching_host
