+++ cmd/root/bundle.go
@@ -2,16 +2,12 @@ package root
 
 import (
 	"context"
-	"fmt"
 
 	"github.com/databricks/cli/bundle"
 	"github.com/databricks/cli/bundle/env"
 	"github.com/databricks/cli/bundle/phases"
-	"github.com/databricks/cli/libs/auth"
 	"github.com/databricks/cli/libs/cmdctx"
 	"github.com/databricks/cli/libs/diag"
-	"github.com/databricks/cli/libs/dyn"
-	"github.com/databricks/cli/libs/dyn/dynvar"
 	envlib "github.com/databricks/cli/libs/env"
 	"github.com/spf13/cobra"
 	"golang.org/x/exp/maps"
@@ -97,8 +93,6 @@ func configureBundle(cmd *cobra.Command, b *bundle.Bundle) (*bundle.Bundle, diag
 	if diags.HasError() {
 		return b, diags
 	}
-
-	diags = diags.Extend(checkVariablesInAuthFields(b))
 
 	// Set the auth configuration in the command context. This can be used
 	// downstream to initialize a API client.
@@ -186,65 +180,4 @@ func initEnvironmentFlag(cmd *cobra.Command) {
 	cmd.PersistentFlags().StringP("environment", "e", "", "bundle target to use (if applicable)")
 	cmd.PersistentFlags().MarkDeprecated("environment", "use --target flag instead")
 	cmd.RegisterFlagCompletionFunc("environment", targetCompletion)
-}
-
-func checkVariablesInAuthFields(b *bundle.Bundle) diag.Diagnostics {
-	authFields := []string{
-		// Generic attributes.
-		"host",
-		"profile",
-		"auth_type",
-		"metadata_service_url",
-
-		// OAuth specific attributes.
-		"client_id",
-
-		// Google specific attributes.
-		"google_service_account",
-
-		// Azure specific attributes.
-		"azure_resource_id",
-		"azure_use_msi",
-		"azure_client_id",
-		"azure_tenant_id",
-		"azure_environment",
-		"azure_login_app_id",
-	}
-
-	diags := diag.Diagnostics{}
-
-	for _, fieldName := range authFields {
-		p := dyn.NewPath(dyn.Key("workspace"), dyn.Key(fieldName))
-		v, err := dyn.GetByPath(b.Config.Value(), p)
-		if dyn.IsNoSuchKeyError(err) {
-			continue
-		}
-		if err != nil {
-			return diag.FromErr(err)
-		}
-
-		vv, ok := v.AsString()
-		if !ok {
-			continue
-		}
-
-		// Check if the field contains interpolation.
-		if dynvar.ContainsVariableReference(vv) {
-			envVar, ok := auth.GetEnvFor(fieldName)
-			if !ok {
-				continue
-			}
-
-			diags = append(diags, diag.Diagnostic{
-				Severity: diag.Warning,
-				Summary:  "Variable interpolation is not supported for fields that configure authentication",
-				Detail: fmt.Sprintf(`Interpolation is not supported for the field %s. Please set
-the %s environment variable if you wish to configure this field at runtime.`, p.String(), envVar),
-				Locations: v.Locations(),
-				Paths:     []dyn.Path{p},
-			})
-		}
-	}
-
-	return diags
 }
