+++ libs/databrickscfg/loader.go
@@ -8,7 +8,6 @@ import (
 	"strings"
 
 	"github.com/databricks/cli/libs/log"
-	"github.com/databricks/cli/libs/workspace"
 	"github.com/databricks/databricks-sdk-go/config"
 	"gopkg.in/ini.v1"
 )
@@ -75,7 +74,7 @@ func (l profileFromHostLoader) Configure(cfg *config.Config) error {
 		return fmt.Errorf("cannot parse config file: %w", err)
 	}
 	// Normalized version of the configured host.
-	host := workspace.NormalizeHost(cfg.Host)
+	host := normalizeHost(cfg.Host)
 	match, err := findMatchingProfile(configFile, func(s *ini.Section) bool {
 		key, err := s.GetKey("host")
 		if err != nil {
@@ -84,7 +83,7 @@ func (l profileFromHostLoader) Configure(cfg *config.Config) error {
 		}
 
 		// Check if this section matches the normalized host
-		return workspace.NormalizeHost(key.Value()) == host
+		return normalizeHost(key.Value()) == host
 	})
 	if err == errNoMatchingProfiles {
 		return nil
