+++ libs/databrickscfg/ops.go
@@ -9,7 +9,6 @@ import (
 	"strings"
 
 	"github.com/databricks/cli/libs/log"
-	"github.com/databricks/cli/libs/workspace"
 	"github.com/databricks/databricks-sdk-go/config"
 	"gopkg.in/ini.v1"
 )
@@ -68,7 +67,7 @@ func matchOrCreateSection(ctx context.Context, configFile *config.File, cfg *con
 			return false
 		}
 		// Check if this section matches the normalized host
-		return workspace.MatchHost(host, cfg.Host)
+		return normalizeHost(host) == normalizeHost(cfg.Host)
 	})
 	if err == errNoMatchingProfiles {
 		section, err = configFile.NewSection(cfg.Profile)
@@ -139,7 +138,7 @@ func ValidateConfigAndProfileHost(cfg *config.Config, profile string) error {
 	}
 
 	// Normalized version of the configured host.
-	host := workspace.NormalizeHost(cfg.Host)
+	host := normalizeHost(cfg.Host)
 	match, err := findMatchingProfile(configFile, func(s *ini.Section) bool {
 		return profile == s.Name()
 	})
@@ -147,8 +146,8 @@ func ValidateConfigAndProfileHost(cfg *config.Config, profile string) error {
 		return err
 	}
 
-	hostFromProfile := workspace.NormalizeHost(match.Key("host").Value())
-	if hostFromProfile != "" && host != "" && !workspace.MatchHost(hostFromProfile, host) {
+	hostFromProfile := normalizeHost(match.Key("host").Value())
+	if hostFromProfile != "" && host != "" && hostFromProfile != host {
 		return fmt.Errorf("config host mismatch: profile uses host %s, but CLI configured to use %s", hostFromProfile, host)
 	}
